+++ indico/util/marshmallow.py
@@ -23,7 +23,7 @@ from indico.core.permissions import get_unified_permissions
 from indico.util.date_time import now_utc
 from indico.util.i18n import _
 from indico.util.placeholders import get_missing_placeholders
-from indico.util.string import get_format_placeholders, has_relative_links
+from indico.util.string import get_format_placeholders, has_endpoint_links, has_relative_links
 from indico.util.user import principal_from_identifier
 
 
@@ -60,6 +60,24 @@ def no_relative_urls(value):
     """Validator that checks links/images use absolute URLs."""
     if value and has_relative_links(value):
         raise ValidationError(_('Links and images may not use relative URLs.'))
+
+
+def no_endpoint_links(endpoint, query_args=frozenset()):
+    """Validator that rejects links pointing to the given endpoint.
+
+    This checks only ``a[href]``, but not URLs present as plain text or in any
+    other (unexpected) places.
+
+    It is intended for places where a placeholder containing a unique/secret link
+    is expected to be used, but users may copy&paste emails from the event log or
+    other places and accidentally include a hardcoded token (of someone else).
+    """
+
+    def _validator(value):
+        if has_endpoint_links(value, endpoint, query_args):
+            raise ValidationError(_('This field contains a forbidden link. Did you mean to use a placeholder?'))
+
+    return _validator
 
 
 def validate_format_placeholders(format_string, valid_placeholders, required_placeholders=None):