+++ CHANGES.rst
@@ -11,11 +11,17 @@ Improvements
 ^^^^^^^^^^^^
 
 - Disallow comments/judgments on outdated editables (:pr:`7067`)
+- Log original email content (with placeholders) when emailing registrants or
+  sending invitations (:pr:`7093`)
+- Disallow sending registration emails or invitations containing hardcoded (and
+  usually incorrect) token links (:pr:`7093`)
 
 Bugfixes
 ^^^^^^^^
 
-- Nothing so far :)
+- Do not allow sending registration invitation reminders without the invitation
+  link placeholder (:pr:`7093`)
+- Correctly log the user sending a registration invitation reminder (:pr:`7093`)
 
 Accessibility
 ^^^^^^^^^^^^^